Correct the logic around SER vs DFR modes
SER/DFR' does not in fact switch between the internal and external
VREFs for ADC conversion (that's just PD1). SER indicates to use the
internal OR external VREF for touchscreen coordinate reads, whichever is
provided. DFR indicates to use the voltage across the touchscreen
drivers for the reference voltage for touchscreen reads, which we pretty
much always want.