Remove == and != operator overloads
A comparison that also compares the Z-axis/pressure would result in a
floating point comparison, and including touch pressure in the
comparison of two points may or may not be what the user wants (both
options might be surprising behavior). For now it's probably better for
the user to compare coordinates manually if they want to do that.